refactor: util 제거 및 dto 추가
