+++ src/main/java/chess/view/OutputView.java
@@ -1,5 +1,6 @@
 package chess.view;
 
+import chess.dto.ChessBoardDto;
 import java.util.List;
 
 public class OutputView {
@@ -8,7 +9,7 @@ public class OutputView {
     private static final String GAME_COMMAND_MOVE_DESCRIPTION = String.format(
         "%s source위치 target위치 - 예. %s b2 b3", Command.MOVE.getAnswer(), Command.MOVE.getAnswer());
     private static final String GAME_COMMAND_REQUEST = String.format(
-        "> 게임 시작: %s\n> 게임 종료: %s\n> 게임 이동 :%s\n",
+        "> 게임 시작 : %s\n> 게임 종료 : %s\n> 게임 이동 : %s\n",
         Command.START.getAnswer(), Command.END.getAnswer(), GAME_COMMAND_MOVE_DESCRIPTION);
 
     public void printStartMessage() {
@@ -16,8 +17,9 @@ public class OutputView {
         System.out.println(GAME_COMMAND_REQUEST);
     }
 
-    public void printBoard(List<List<String>> board) {
+    public void printBoard(final ChessBoardDto chessBoard) {
         System.out.println();
+        List<List<String>> board = chessBoard.getBoard();
         for (List<String> rank : board) {
             printRank(rank);
         }
